# API_Documentation.yaml
openapi: '3.0.2'
info:
  title: ISS Capstone API
  version: '1.0'
  description: 
    This is the API documentation for the ISS Capstone project.
    For more information, contact the developers.
servers:
  - url: https://api.server.test/v1
paths:
...
  /Items:
    get:
      summary: Get all items
      tags:
        - Items
      responses:
        '200':
          description: Returns all items in the data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties: 
                    _id:
                      type: integer
                      example: 1
                    Title:
                      type: string
                      example: "Item 1"
                    Description:
                      type: string
                      example: "This is item 1"
                    Price:
                      type: integer
                      example: 100
                    ImageURL:
                      type: string
                      example: "localhost:8080/image1"
                    CategoryId:
                      type: integer
                      example: 1
  
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true

        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true

      security:
        - bearerAuth: []
  /Items/{_id}:
    get:
      summary: Get item by ID
      tags:
        - Items
      parameters:
        - in: path
          name: _id
          schema:
            type: integer
          required: true
          description: Numeric ID of the item to get
      responses:
        '200':
          description: Get all registered items
          content:
            application/json:
              schema:
                type: object
                example:
                    _id: 1
                    Title: "Item 1"
                    Description: "This is item 1"
                    Price: 100
                    ImageURL: "localhost:8080/image1"
                    CategoryId: 1
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true

        "404":
          description: "Item not found"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true

        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true

      security:
        - bearerAuth: []
  /Items/create:
    post:
      tags:
      - Items
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Title:
                  type: string
                  example: "Item 1"
                Description:
                  type: string
                  example: "This is item 1"
                Price:
                  type: integer
                  example: 100
                ImageURL:
                  type: string
                  example: "localhost:8080/image1"
                CategoryId:
                  type: integer
                  example: 1
              required:
                - Title
                - Description
                - Price
                - ImageURL
                - CategoryId
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: false
                  message: "Item created successfully!"
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true

        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true

        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true

      security:
          - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT